Move errors to internal to prevent external libraries from triggering plugin removal
